move LogEventTracingProperties into Enrichers namespace
